+++ torch_spyre/csrc/module.h
@@ -28,6 +28,7 @@ struct SharedOwnerCtx {
   flex::DeviceMemoryAllocationPtr owner;
   signed char device_id;
   size_t nbytes;
+  c10::Device device;
 };
 
 class GlobalRuntime {
