+++ torch_spyre/csrc/spyre_mem_stat.h
@@ -1,0 +1,34 @@
+/*
+ * Copyright 2025 The Torch-Spyre Authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include <c10/core/Allocator.h>
+
+struct DeviceStats {
+  c10::CachingAllocator::StatArray allocation;
+  c10::CachingAllocator::StatArray allocated_bytes;
+};
+
+namespace spyre {
+
+const ::DeviceStats& get_stats(std::optional<int> device_index);
+
+void reset_peak_stats(std::optional<int> device_index);
+
+void reset_accumulated_stats(std::optional<int> device_index);
+
+}  // namespace spyre
